Moved _azure module to azure
