+++ python/packages/autogen-ext/src/autogen_ext/models/azure/__init__.py
@@ -1,0 +1,5 @@
+from ._azure_ai_client import AzureAIChatCompletionClient
+
+__all__ = [
+    "AzureAIChatCompletionClient"
+]
